Added Classes for Meals and Ingredients
Created classes for meals and ingredients, ingredients take in an object which reflects the ingredients macro nutrients, per x number of grams and contains ingredient's usage grams. Meals contain an array of ingredients which can be added and set, Ingredient class has a function for calculating macros for an ingredient and Meals have a function which takes each ingredient's macro nutrient counts and sums them up thus showing the macros for the specific meal with ingredients. Added a test.html to test classes work, DailyIntake will have an array of meals and a function which calculates the sum of the meal's macros and returns the macros for the day. Will maybe add a way of serialising DailyIntake to allow user's to keep their days stored and maybe do the same for Meals + Ingredients. This serialised version of the objects could be potentially stored in LocalStorage or cookies which can then be accessed via JS in the browser and turned back into the objects they once were.